First commit
Been working on this for few months now, the point was to start (re)learning my C# coding and coding in general. Order execution was working in the first version when I was testing my idea, just had to start everything from scratch when I started working on UI and adding too much unimportant features (for now).